requirements.txt
imageio>=2.26
matplotlib>=3.7
numpy>=1.24
Pillow>=9.0
